No more messages for trying to register an already registered factory
iPOPO bundles are no more started unless it is asked, so iPOPO does not
try any more to register a factory we have already registered
programmatically.
