+++ src/fastoad/module_management/exceptions.py
@@ -1,0 +1,27 @@
+"""
+Exceptions for module_management package
+"""
+#  This file is part of FAST : A framework for rapid Overall Aircraft Design
+#  Copyright (C) 2019  ONERA/ISAE
+#  FAST is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+from fastoad.exceptions import FastError
+
+
+class FastDuplicateFactoryError(FastError):
+    """
+    Raised when trying to register a factory with an already used name
+    """
+
+    def __init__(self, factory_name):
+        super().__init__('Name "%s" is already used.' % factory_name)
+        self.factory_name = factory_name
